requirements.txt
python-dotenv>=1.0.1
pyyaml>=6.0.2
crewai>=0.51.0


